Refactor PresetOverviewScreen to separate module
